checkpoint before vacation - important: added custom v0 finder based on standard ALICE offline v0 finder - small changes: Radius() -> Radius2D(), added DoFit(), PdgCode made int - added cut flow hists - CMake: export compile commands true by default, added ROOT::Hist - clang: removed boost checks - updated README accordingly

+++ include/Math/Constants.hxx
@@ -32,7 +32,7 @@ static constexpr std::string_view PiMinus{"PM"};
 static constexpr std::string_view PiPlus{"PP"};
 }  // namespace Acronym
 
-enum class PdgCode : short {
+enum class PdgCode : int {
     AntiLambda = -3122,
     Lambda = 3122,
     KaonZeroShort = 310,
